requirements.txt
prettytable
colorama
PyInstaller
pyautogui
opencv-python
pillow
numpy